+++ changelog/mediation/mobileads-bigoads/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 4.6.0.5
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.12.3`
+
 ## Version 4.6.0.4
 
 ### Updated