+++ changelog/mediation/mobileads-inmobi/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 10.7.8.9
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.12.3`
+
 ## Version 10.7.8.8
 
 ### Updated